+++ src/providers/MarkdownFoldingProvider.ts
@@ -1,4 +1,5 @@
-import { CancellationToken, FoldingContext, FoldingRange, FoldingRangeKind, FoldingRangeProvider, TextDocument } from 'vscode';
+import { CancellationToken, FoldingContext, FoldingRange, FoldingRangeKind, FoldingRangeProvider, Range, TextDocument, window, TextEditorDecorationType, Position } from 'vscode';
+import { FrontMatterDecorationProvider } from './FrontMatterDecorationProvider';
 
 export class MarkdownFoldingProvider implements FoldingRangeProvider {
 
@@ -16,11 +17,17 @@ export class MarkdownFoldingProvider implements FoldingRangeProvider {
           start = i;
         } else if (start !== null && end === null) {
           end = i;
+
+          const range = new Range(new Position(start, 0), new Position(end, line.length));
+          const decoration = new FrontMatterDecorationProvider().get();
+          window.activeTextEditor?.setDecorations(decoration, [range]);
+
           ranges.push(new FoldingRange(start, end, FoldingRangeKind.Region));
           return ranges;
         }
       }
     }
+
     return ranges;
   }
 }